+++ tests/testthat/test-03-select_individual.R
@@ -1,5 +1,16 @@
 library(mockery)
 
+test_that("select_individual() works with non-existing config_folder", {
+  temp_config_dir <- tempfile("citeme_config")
+  expect_false(dir.exists(temp_config_dir))
+  root <- mock_r_user_dir(temp_config_dir)("citeme", which = "data")
+  stub(select_individual, "R_user_dir", root, depth = 2)
+  expect_error(
+    select_individual(lang = "en-GB"),
+    "No available individuals in a non-interactive session."
+  )
+})
+
 test_that("select_individual() returns correct individual", {
   new_person <- data.frame(
     given = "Given",