+++ NEWS.md
@@ -1,3 +1,9 @@
+# citeme 0.1.4
+
+* Add `ask_keywords()` function for interactive keyword input.
+* Handle non-existing configuration directory (#15 by @damianoodolni).
+* Make sure to use `true` and `false` when writing logical values to YAML.
+
 # citeme 0.1.3
 
 * Handle `README` without language badge.